CPA

Intros CPA

Power estimation function

Intros concept power estimation function. Some examples?

CPA attacks against AES

CryptoSideChannel.CPA.CPA_AES_analyze — Function

CPA_AES_analyze(sample_function, power_estimate)

Performs a CPA attack against AES, where traces are collected from a specified function.

Arguments

sample_function: single-argument function that takes an input AES input (MVector{16, UInt8}) and returns a power trace as an array of numbers for this input.

power_estimate: a function returning an estimate for power consumption, given a plaintext, and a key guess at the position key_guess_index. The signature should be compatible to hypothesis(plaintext::MVector{16, UInt8, key_guess_index::Int, key_guess::Int) For example, a classical implementation of this function for AES with Hamming weight would be julia power_estimate(plaintext, key_guess_index, key_guess) = Base.count_ones(AES.c_sbox[(plaintext[key_guess_index] ⊻ key_guess)+1])

CryptoSideChannel.CPA.CPA_AES_analyze_traces — Function

CPA_AES_analyze_traces(plaintexts::Vector, traces::Matrix, power_estimate)

Performs a CPA attack against AES on given traces.

Arguments

plaintexts: A vector of size N, where N is the number of power traces sampled.

traces: A matrix of size M * N, where M is the number of samples per trace. Power traces are stored in column-major order, i.e. it is expected that traces[i,:] refers to the powertrace generated with plaintexts[i]

power_estimate: A function that takes a plaintext::MVector{16, UInt6}, a key index (1 <= key_guess_index <= 16), and a key guess (0 <= key_guess <= 255) and returns an hypothesis on power consumption. For example, a classical implementation of this function for AES with Hamming weight would be power_estimate(plaintext, key_guess_index, key_guess) = Base.count_ones(AES.c_sbox[(plaintext[key_guess_index] ⊻ key_guess)+1])

CPA attacks against SPECK

Attacks against SPECK are a bit more difficult, since both key parts (the left 64 bits and the right 64 bits) have to be attacked differently.

CryptoSideChannel.CPA.CPA_SPECK_analyze — Function

CPA_SPECK_analyze(sample_function)

Performs a CPA attack against SPECK.

Arguments

sample_function: a single-argument function that takes a SPECK input (Tuple{UInt64, UInt64}) and returns a power trace (array of numbers) for this input.

N: the amount of traces to collect

Returns

The reconstructed SPECK key as a Tuple{UInt64, UInt64}

CryptoSideChannel.CPA.CPA_SPECK_analyze_traces — Function

CPA_SPECK_analyze_traces(plaintexts::Vector, traces::Matrix)

Perform a CPA attack against SPECK on the provided traces.

Arguments

plaintexts: A Vector of size N, where N is the number of power traces sampled.

traces: A Matrix of size M * N, where M is the number of samples per trace.

Power traces are stored in column-major order, i.e. it is expected that traces[i,:] refers to the powertrace generated with plaintexts[i]

Returns

The reconstructed SPECK key as a Tuple{UInt64, UInt64}

Real-world attacks against AES

The module CryptoSideChannel.AES.AES_RealWorld implements a real-world CPA attack against the Northeastern University TeSCASE dataset, which can be obtained here. The attacked traces have been generated on a SASEBO board that executed the AES algorithm.

First, a power estimate for the real-world data has to be found. The paper "Scalable and efficient implementation of correlation power analysis using graphics processing units" suggests to use a Hamming distance estimation that targets the last round of AES. Roughly following the proposed algorithm in this paper, our implementation uses the following method to estimate power consumption:

CryptoSideChannel.CPA.AES_RealWorld.hamming_distance_power_estimate — Function

hamming_distance_power_estimate(ciphertext, key_guess_index, key_guess)

A power estimation function targeting the last round of AES under the Hamming distance model, suitable for the real-word data provided. For more information on power estimation, see the main section.