Template

CryptoSideChannel.TemplateAttacks.Template — Type

The Template struct stores a noise distribution, as well as values for integers that are logged.

If the integer x is logged, a random vector from distribution is drawn. Then, values[x] is added to this random vector.

CryptoSideChannel.TemplateAttacks.random_diagonal_multivariate_distribution — Function

random_diagonal_multivariate_distribution(d::Integer)

Generate a random normal multivariate distribution over d dimensions. The mean will be a vector chosen randomly from $\[0, 1\]^d$.

Note that all random values are uncorrelated. Hence, our covariance matrix is a diagonal matrix.

CryptoSideChannel.TemplateAttacks.random_uncorrelated_template — Function

random_uncorrelated_template(dimensions::Integer, max_value::Integer)

Generate a random template over d dimensions, that supports operations on integers between $0$ and max_value.

CryptoSideChannel.TemplateAttacks.LikelyKey — Type

This struct merges different key bytes for which probabilities are known to a whole key, by iterating first over keys that are more likely.

Keys are stored as lists of lists, where the outer lists represent the respective key byte (i.e. the first list represents the first key byte). The inner lists must be sorted according to the probability of a specific byte occuring.